GUI/requirements.txt
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
contextlib2==0.6.0.post1
execnet==1.7.1
future==0.18.2
importlib-metadata==1.6.1
mock==4.0.2
more-itertools==8.4.0
mysql-connector-python==8.0.20
numpy==1.18.5
opencv-python==4.2.0.34
packaging==20.4
path==13.1.0
path.py==12.4.0
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.12.2
py==1.8.2
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pytest==5.4.3
pytest-shutil==1.7.0
PyYAML==5.3.1
six==1.15.0
termcolor==1.1.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
wcwidth==0.2.4
XlsxWriter==1.2.9
zipp==3.1.0
